refactor: auth and user module and related import statements
